+++ workalendar/tests/test_registry_europe.py
@@ -13,7 +13,7 @@ from ..europe import (
 # Spain & regions
 from ..europe import (
     Spain, Andalusia, Aragon, Catalonia, CastileAndLeon, CastillaLaMancha,
-    CanaryIslands, Extremadura, Galicia
+    CanaryIslands, Extremadura, Galicia, BalearicIslands
 )
 
 # Switzerland
@@ -54,7 +54,7 @@ SWITZERLAND_REGION_CLASSES = (
 
 SPAIN_REGION_CLASSES = (
     Andalusia, Aragon, Catalonia, CastileAndLeon, CastillaLaMancha,
-    CanaryIslands, Extremadura, Galicia
+    CanaryIslands, Extremadura, Galicia, BalearicIslands
 )